requirements.txt
fastapi>=0.103.1
psycopg2-binary>=2.9.7
sqlmodel>=0.0.8
uvicorn>=0.23.2
python-multipart==0.0.6
python-jose[cryptography]
passlib[bcrypt]
python-dotenv==1.0.0